feat: Account Configuration Load & Store
Each backend can store the account's information in a abstract json
Value which will then be stored encrypted.

This patch also introduces the AuthRefresher which will be used
to re-authenticate existing sessions.

The only thing remaining is how to handle accounts that are logged out
and may not have a valid access token anymore.