refactor: New architecture
In the previous iteration, all of the polling and account management was
performed on separate thread.

In this new version the management is performed on the calling thread
and polling can be optionally performed on the calling thread or a
dedicated thread.

Finally, the observer no longer performs auto-polling, this allows for
more flexibility when integrating in other systems/applications.