refactor: Refactor old system
The previous version required an active instance present at runtime
to keep everything working. In this refactored version data data is
always read from the database. This makes it easier to integrate in
Android later.

Additionally, account login is now separate to accommodate peculiarities
of other possible backends. The result is a greatly simplified backend
interface.