fix: Fix poll intervals not triggering for < 15 min
The bug was caused by using the `ExstingWorkPolicy.KEEP` when scheduling
the next one time worker, it would cancel out the follow up job. Since
the job that schedules the follow up has the same TAG and is still
running when the follow up is scheduled, `WorkManager` would not record
the new job causing future runs to never be scheduled.